restructed file tree and renamed files
