+++ src/pages/CohortAnalyzer/CohortAnalyzer.js
@@ -12,14 +12,14 @@ import DownloadSelectedCohort from "./downloadCohort/DownloadSelectedCohorts";
 import client from "../../utils/graphqlClient";
 import ToolTip from "@bento-core/tool-tip/dist/ToolTip";
 import Stats from '../../components/Stats/GlobalStatsController';
-import DeleteConfirmationModal from "../inventory/cohortModal/components/deleteConfirmationModal";
+import DeleteConfirmationModal from "../inventory/CohortModal/components/shared/DeleteConfirmationModal";
 import NavigateAwayModal from './navigateAwayModal';
 import sortIcon from "../../assets/icons/sort_icon.svg";
 import placeHolder from "../../assets/vennDigram/placeHolder.png";
 import ChartVenn from "./vennDiagram/ChartVenn";
 import CheckBoxCustom from "./customCheckbox/CustomCheckbox";
-import { CohortModalContext } from "../inventory/cohortModal/CohortModalContext";
-import CohortModal from "../inventory/cohortModal/CohortModal";
+import { CohortModalContext } from "../inventory/CohortModal/CohortModalContext";
+import CohortModal from "../inventory/CohortModal/CohortModal";
 import Alert from '@material-ui/lab/Alert';
 import { useGlobal } from "../../components/Global/GlobalProvider";
 import questionIcon from "../../assets/icons/Question_icon_2.svg";
