+++ src/pages/CohortAnalyzer/CohortAnalyzerController.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { CohortStateProvider } from "../../components/CohortSelectorState/CohortStateContext";
 import { CohortAnalyzer } from "./CohortAnalyzer";
-import { CohortModalProvider } from "../inventory/cohortModal/CohortModalContext";
+import { CohortModalProvider } from "../inventory/CohortModal/CohortModalContext";
 
 const CohortAnalyzerController = (state) => {
   return (
