+++ src/pages/inventory/cohortModal/components/CohortDetails/components/CohortMetadata.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect, useRef, useCallback, useMemo, useContext, memo } from 'react';
 import { withStyles } from '@material-ui/core';
-import DEFAULT_CONFIG from '../config';
-import { CohortModalContext } from '../CohortModalContext';
-import { CohortStateContext } from '../../../../components/CohortSelectorState/CohortStateContext';
+import DEFAULT_CONFIG from '../../../config';
+import { CohortModalContext } from '../../../CohortModalContext';
+import { CohortStateContext } from '../../../../../../components/CohortSelectorState/CohortStateContext';
 
 const CohortMetadata = (props) => {
     const { config, classes } = props;
