+++ src/pages/inventory/cohortModal/components/CohortDetails/components/ParticipantList/ParticipantList.js
@@ -1,10 +1,10 @@
 import React, { useState, useContext, useCallback, useEffect, useMemo, memo } from 'react';
 import { withStyles, Button } from '@material-ui/core';
-import { CohortStateContext } from '../../../../components/CohortSelectorState/CohortStateContext.js';
-import { CohortModalContext } from '../CohortModalContext.js';
-import SearchBar from './searchBar';
-import ParticipantTable from './participantTable';
-import DEFAULT_CONFIG from '../config';
+import { CohortStateContext } from '../../../../../../../components/CohortSelectorState/CohortStateContext.js';
+import { CohortModalContext } from '../../../../CohortModalContext.js';
+import SearchBar from './components/SearchBar';
+import ParticipantTable from './components/ParticipantTable';
+import DEFAULT_CONFIG from '../../../../config';
 
 const ParticipantList = (props) => {
     const { classes, localCohort, setLocalCohort, handleSave, closeModal, config } = props;
