+++ src/pages/inventory/cohortModal/components/CohortDetails/components/ParticipantList/components/ParticipantTable.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useRef, useContext, useCallback, useMemo, memo } from 'react';
 import { withStyles } from '@material-ui/core';
-import { CohortModalContext } from '../CohortModalContext.js';
-import { deletionTypes } from './deleteConfirmationModal';
-import TrashCanIconBlue from '../../../../assets/icons/Trash_Can_Icon_Blue.svg';
-import TrashCanIconRed from '../../../../assets/icons/Trash_Can_Icon_Red.svg';
-import SortingIcon from '../../../../assets/icons/Sorting_Icon.svg';
+import { CohortModalContext } from '../../../../../CohortModalContext.js';
+import { deletionTypes } from '../../../../shared/DeleteConfirmationModal';
+import TrashCanIconBlue from '../../../../../../../../assets/icons/Trash_Can_Icon_Blue.svg';
+import TrashCanIconRed from '../../../../../../../../assets/icons/Trash_Can_Icon_Red.svg';
+import SortingIcon from '../../../../../../../../assets/icons/Sorting_Icon.svg';
 
 // Constants
 const SCROLLBAR_WIDTH = '6px';
