+++ src/pages/inventory/cohortModal/components/CohortDetails/components/ParticipantList/components/index.js
@@ -1,0 +1,2 @@
+export { default as SearchBar } from './SearchBar';
+export { default as ParticipantTable } from './ParticipantTable';
