+++ src/pages/inventory/cohortModal/components/CohortDetails/components/ParticipantList/index.js
@@ -1,0 +1,1 @@
+export { default } from './ParticipantList';
