+++ src/pages/inventory/cohortModal/components/CohortDetails/components/index.js
@@ -1,0 +1,3 @@
+export { default as CohortMetadata } from './CohortMetadata';
+export { default as ActionButtons } from './ActionButtons';
+export { default as ParticipantList } from './ParticipantList';
