+++ src/pages/inventory/cohortModal/components/CohortDetails/index.js
@@ -1,0 +1,1 @@
+export { default } from './CohortDetails';
