+++ src/pages/inventory/cohortModal/components/CohortList/index.js
@@ -1,0 +1,2 @@
+// CohortList feature components
+export { default } from './CohortList';
