+++ src/pages/inventory/cohortModal/components/index.js
@@ -1,0 +1,3 @@
+export { default as CohortList } from './CohortList';
+export { default as CohortDetails } from './CohortDetails';
+export * from './shared';
