+++ src/pages/inventory/cohortModal/components/shared/DeleteConfirmationModal.js
