+++ src/pages/inventory/cohortModal/components/shared/index.js
@@ -1,0 +1,3 @@
+// Shared modal components
+export { default as AlertManager } from './AlertManager';
+export { default as DeleteConfirmationModal, deletionTypes } from './DeleteConfirmationModal';
