+++ src/pages/inventory/tabs/wrapperConfig/customButton.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { useGlobal } from '../../../../components/Global/GlobalProvider';
 import { onCreateNewCohort } from '../../../../components/CohortSelectorState/store/action';
 import { CohortStateContext } from '../../../../components/CohortSelectorState/CohortStateContext';
-import { CohortModalContext } from '../../cohortModal/CohortModalContext';
+import { CohortModalContext } from '../../CohortModal/CohortModalContext';
 import { onRowSelectHidden } from '@bento-core/paginated-table/dist/table/state/Actions';
 
 const ButtonContainer = styled.div`
